+++ packages/stylelint-config-wordpress/scss.js
@@ -11,7 +11,7 @@ module.exports = {
 
     // stylelint-config-wordpress css overrides
     "at-rule-no-unknown": [ true, {
-      ignoreAtRules: [ "extend", "at-root", "debug", "warn", "error", "if", "else", "for", "each", "while", "mixin", "include", "content", "return", "function" ],
+      ignoreAtRules: [ "extend", "at-root", "warn", "error", "if", "else", "for", "each", "while", "mixin", "include", "content", "return", "function" ],
     } ],
   },
 }